+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingBreakdownToken/proposalVotingBreakdownToken.stories.tsx
@@ -9,7 +9,7 @@ const meta: Meta<typeof ProposalVoting.BreakdownToken> = {
     parameters: {
         design: {
             type: 'figma',
-            url: 'https://www.figma.com/design/ISSDryshtEpB7SUSdNqAcw/Aragon-ODS?node-id=16752-18392&m=dev',
+            url: 'https://www.figma.com/design/ISSDryshtEpB7SUSdNqAcw/Governance-UI-Kit?node-id=16752-18392&m=dev',
         },
     },
 };